Add files via upload
Hi!

This is an animation for a running app I made. The general idea is that users will be able to see a small representation of their run while viewing their statistics. I utilized Figma smart animate to move the pin, symbolizing the person on their run, along the path.